create Home component for home screen
