+++ src/App.css
@@ -1,18 +1,3 @@
-.App-header {
-  background-color: green;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
 h1 {
   font-size: 70px;
   text-align: center;
@@ -32,8 +17,4 @@ p, li {
   line-height: 2;
   letter-spacing: 1px;
   margin: 20px;
-}
-
-.link-box {
-  display: flex;
 }
