+++ src/Pages/Home/index.css
@@ -1,0 +1,14 @@
+.home-header {
+    background-color: green;
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    font-size: calc(10px + 2vmin);
+    color: white;
+}
+
+.link-box {
+    display: flex;
+}
